+++ CHANGELOG.md
@@ -2,12 +2,16 @@
 
 ## TODO
 
-- [ ] add contact section
+- [ ] add contact section (emailJs)
 - [ ] find suitable BG to hero section
 - [ ] add [Qualifications](https://codepen.io/xichen/pen/wzpeMM) section Education & Work Experience as timeline (date - title - info)
 - [ ] apply [Alphabet soup](https://codepen.io/mohmmedtaysser/pen/rNQYVwg?editors=1010) effect on hero section
 
 ## History
+
+### 23 August 2023
+
+- add takenshot project
 
 ### 9 August 2023